Rename: 컴포넌트에서  styled-components 분리
