+++ frontend/src/components/Detail/Calendar/index.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react';
-import styled, { css } from 'styled-components';
 import { format, startOfMonth, endOfMonth, startOfWeek, endOfWeek, addDays, isSameMonth } from 'date-fns';
+import { Container, Month, Week, Days, OneWeek, Day } from './index.style';
 
 const Calendar = () => {
   const curDate = new Date();
@@ -53,68 +53,3 @@ const Calendar = () => {
 };
 
 export default Calendar;
-
-const FontCss = css`
-  text-align: left;
-  font-style: normal;
-  font-variant: normal;
-  font-weight: medium;
-  letter-spacing: 0px;
-  opacity: 1;
-`;
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 400px;
-  height: 350px;
-`;
-const Month = styled.div`
-  margin-bottom: 25px;
-  font-size: 20px;
-  line-height: 26px;
-  color: #222222;
-`;
-
-const Week = styled.div`
-  display: flex;
-  margin-bottom: 20px;
-  ${FontCss}
-  font-size: 16px;
-  line-height: 26px;
-  color: #222222;
-  opacity: 1;
-  & > :first-child {
-    color: #e14d4d;
-  }
-  & > :last-child {
-    color: #372cdf;
-  }
-  & > div {
-    width: 100%;
-  }
-`;
-
-const Days = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-const OneWeek = styled.div`
-  display: flex;
-  margin-bottom: 30px;
-  ${FontCss}
-  font-size: 14px;
-  line-height: 26px;
-  color: #888888;
-  & > :first-child {
-    color: #e14d4d;
-  }
-  & > :last-child {
-    color: #372cdf;
-  }
-`;
-
-const Day = styled.div`
-  width: 100%;
-`;
