+++ frontend/src/components/Detail/KakaoMap/index.tsx
